Delete now works on all three entities, from child entity to parent and vice versa.

+++ src/main/java/com/example/adventure/customer/controller/CustomerController.java
@@ -48,9 +48,12 @@ public class CustomerController {
                 new RuntimeException("Not found")));
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/delete/{id}")
     public ResponseEntity<Customer> delete(@PathVariable("id") Long id){
-        return ResponseEntity.ok().body(service.delete(id));
+        if(id != null){
+            return ResponseEntity.ok().body(service.delete(id));
+        }
+        return ResponseEntity.badRequest().build();
     }
 
 }
